requirements.txt
Flask_Cors==3.0.10
requests==2.25.1
pytest==6.2.2
Werkzeug==1.0.1
Flask==1.1.2
flask_mail==0.9.1
Pillow==8.2.0
PyJWT==2.0.1
